Refactor code and add logic to maintain task status on edit
The delete task logic has been separated into deleteTask and removeTask. This allows functions to remove Tasks from the DOM/disk without having to follow the API used by deleteTask. deleteTask is now used only when users click the "X" beside a task to delete.

The modal for editing/creating tasks is more "intelligent" in that it keeps track of the task if being edited. If a task is being created from scratch (i.e. not edited), a new ID is generated instead of looking for the ID on the board.